perf: Task import with Project (#17496)
When tasks are imported with their Project set, import slows down as
number of tasks increase. This is largely due to `load_tasks` being
called in Project which is only required when Form loads. So, we can
skip it during import.

Also, converted some get_doc calls to get_cached_doc.